Remove StripeIntegration.js (already handld in separate directory
Previously, code to get Stripe payment information via webhooks was in this diretory. We have moved all that code to its own sub-directory, which at this point is called instant-currency-event-handlers. That directory also has code to update the top currency pair rates every day. This directory, currently called instant-currency-add-on, will need updated code to check whether the logged in user has paid for Instant Currency Tools for Sheets. There is code to do that in SubscriptionManager.gs/.js, but it will need to be updated.